Modularize sidebar timeline
Separate bucket rendering, timeline controls, window derivation, and DOM interaction helpers behind the existing TimelineView facade while preserving rendering, timing, and translated copy.
